# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Education API System",
        "description": "教育管理系统API文档",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.localhost",
            "description": "开发环境"
        }
    ],
    "paths": {
...
        "/api/invoices": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "获取账单列表",
                "operationId": "0b8e5b661b472432febe76177e627e86",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "页码",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "每页数量",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "账单状态",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "student_id",
                        "in": "query",
                        "description": "学生ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Invoices"
                ],
                "summary": "创建账单",
                "operationId": "9a379bde2decc411b48e2fdc02a315ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "student_id": {
                                        "description": "学生ID",
                                        "type": "integer"
                                    },
                                    "course_id": {
                                        "description": "课程ID",
                                        "type": "integer"
                                    },
                                    "amount": {
                                        "description": "账单金额",
                                        "type": "number"
                                    },
                                    "description": {
                                        "description": "账单描述",
                                        "type": "string"
                                    },
                                    "year_month": {
                                        "description": "年月(YYYYMM)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/invoices/{id}": {
            "get": {
                "tags": [
                    "Invoices"
                ],
                "summary": "获取账单详情",
                "operationId": "6f9874ec6f7301133948074ff6108fcb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "账单ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Invoices"
                ],
                "summary": "更新账单",
                "operationId": "f72d18e23867bcd77416956363bfc090",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "账单ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "description": "账单金额",
                                        "type": "number"
                                    },
                                    "description": {
                                        "description": "账单描述",
                                        "type": "string"
                                    },
                                    "year_month": {
                                        "description": "年月(YYYYMM)",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invoices"
                ],
                "summary": "删除账单",
                "operationId": "f6705ba1adcf67585cb5f13627c592ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "账单ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearer_token": []
        }
    ]
}